Created new component - Mosquitto Auth Server
This step is result of detaching mosquitto authentication API
from piot-server. One of the reasons was problematic startup
of piot-server as it was acting in two roles - auth server
and subscriber to mqtt topics (which needs authorization).
Such combination didn't work well especially in case we would
like to avild of static mosquitto users (hardcoded in file system
password files)